bug hunging

+++ ft_strnstr.c
@@ -28,7 +28,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
     return (NULL);
   if (needle == haystack)
     return ((char *)haystack);
-	while ((len - i) >= ft_strlen(needle) && *haystack)
+	while ((int)((int)len - i) >= (int)ft_strlen(needle) && *haystack)
 	{
 		cmp = (char *)haystack;
 		tmp = (char *)needle;
